+++ driver-core/src/main/java/com/datastax/driver/core/Cluster.java
@@ -23,6 +23,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.SetMultimap;
+import com.google.common.util.concurrent.ListeningExecutorService;
+import com.google.common.util.concurrent.MoreExecutors;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 
 import org.apache.cassandra.utils.MD5Digest;
@@ -161,6 +163,37 @@ public class Cluster {
      */
     public Metrics getMetrics() {
         return manager.metrics;
+    }
+
+    /**
+     * Registers the provided listener to be notified on hosts
+     * up/down/added/removed events.
+     * <p>
+     * Registering the same listener multiple times is a no-op.
+     * <p>
+     * Note that while {@link LoadBalancingPolicy} implements
+     * {@code Host.StateListener}, the configured load balancy does not
+     * need to (and should not) be registered through this  method to
+     * received host related events.
+     *
+     * @param listener the new {@link Host.StateListener} to register.
+     * @return this {@code Cluster} object;
+     */
+    public Cluster register(Host.StateListener listener) {
+        manager.listeners.add(listener);
+        return this;
+    }
+
+    /**
+     * Unregisters the provided listener from being notified on hosts events.
+     * <p>
+     * This method is a no-op if {@code listener} hadn't previously be
+     * registered against this monitor.
+     *
+     * @param listener the {@link Host.StateListener} to unregister.
+     */
+    public void unregister(Host.StateListener listener) {
+        manager.listeners.remove(listener);
     }
 
     /**
@@ -570,7 +603,7 @@ public class Cluster {
         // applied in the order received.
         final ScheduledExecutorService scheduledTasksExecutor = Executors.newScheduledThreadPool(1, threadFactory("Scheduled Tasks-%d"));
 
-        final ExecutorService executor = Executors.newCachedThreadPool(threadFactory("Cassandra Java Driver worker-%d"));
+        final ListeningExecutorService executor = MoreExecutors.listeningDecorator(Executors.newCachedThreadPool(threadFactory("Cassandra Java Driver worker-%d")));
 
         final AtomicBoolean isShutdown = new AtomicBoolean(false);
 
@@ -579,6 +612,8 @@ public class Cluster {
         // Note: we could move this down to the session level, but since prepared statement are global to a node,
         // this would yield a slightly less clear behavior.
         final Map<MD5Digest, PreparedStatement> preparedQueries = new ConcurrentHashMap<MD5Digest, PreparedStatement>();
+
+        private final Set<Host.StateListener> listeners = new CopyOnWriteArraySet<Host.StateListener>();
 
         private Manager(List<InetAddress> contactPoints, Configuration configuration) {
             logger.debug("Starting new cluster with contact points " + contactPoints);
@@ -674,10 +709,17 @@ public class Cluster {
             if (isShutdown.get())
                 return;
 
+            if (host.isUp())
+                return;
+
+            host.setUp();
+
             // If there is a reconnection attempt scheduled for that node, cancel it
             ScheduledFuture<?> scheduledAttempt = host.reconnectionAttempt.getAndSet(null);
-            if (scheduledAttempt != null)
+            if (scheduledAttempt != null) {
+                logger.debug("Cancelling reconnection attempt since node is UP");
                 scheduledAttempt.cancel(false);
+            }
 
             try {
                 prepareAllQueries(host);
@@ -697,6 +739,9 @@ public class Cluster {
             controlConnection.onUp(host);
             for (Session s : sessions)
                 s.manager.onUp(host);
+
+            for (Host.StateListener listener : listeners)
+                listener.onUp(host);
         }
 
         @Override
@@ -706,10 +751,18 @@ public class Cluster {
             if (isShutdown.get())
                 return;
 
+            if (!host.isUp())
+                return;
+
+            host.setDown();
+
             loadBalancingPolicy().onDown(host);
             controlConnection.onDown(host);
             for (Session s : sessions)
                 s.manager.onDown(host);
+
+            for (Host.StateListener listener : listeners)
+                listener.onDown(host);
 
             // Note: we basically waste the first successful reconnection, but it's probably not a big deal
             logger.debug("{} is down, scheduling connection retries", host);
@@ -721,7 +774,7 @@ public class Cluster {
 
                 protected void onReconnection(Connection connection) {
                     logger.debug("Successful reconnection to {}, setting host UP", host);
-                    host.getMonitor().setUp();
+                    onUp(host);
                 }
 
                 protected boolean onConnectionException(ConnectionException e, long nextDelayMs) {
@@ -756,6 +809,9 @@ public class Cluster {
             controlConnection.onAdd(host);
             for (Session s : sessions)
                 s.manager.onAdd(host);
+
+            for (Host.StateListener listener : listeners)
+                listener.onAdd(host);
         }
 
         @Override
@@ -763,11 +819,27 @@ public class Cluster {
             if (isShutdown.get())
                 return;
 
+            host.setDown();
+
             logger.trace("Removing host {}", host);
             loadBalancingPolicy().onRemove(host);
             controlConnection.onRemove(host);
             for (Session s : sessions)
                 s.manager.onRemove(host);
+
+            for (Host.StateListener listener : listeners)
+                listener.onRemove(host);
+        }
+
+        public boolean signalConnectionFailure(Host host, ConnectionException exception) {
+            // If already down, don't signal again
+            if (!host.isUp())
+                return true;
+
+            boolean isDown = host.signalConnectionFailure(exception);
+            if (isDown)
+                onDown(host);
+            return isDown;
         }
 
         public Host addHost(InetAddress address, boolean signal) {
@@ -954,7 +1026,7 @@ public class Cluster {
                                         // first time we heard about that node apparently, add it
                                         addHost(stc.node.getAddress(), true);
                                     } else {
-                                        hostUp.getMonitor().setUp();
+                                        onUp(hostUp);
                                     }
                                     break;
                                 case DOWN:
@@ -964,7 +1036,7 @@ public class Cluster {
                                     // right away, so we favor the detection to make the Host.isUp method more reliable.
                                     Host hostDown = metadata.getHost(stc.node.getAddress());
                                     if (hostDown != null) {
-                                        hostDown.getMonitor().setDown();
+                                        onDown(hostDown);
                                     }
                                     break;
                             }