+++ driver-core/src/main/java/com/datastax/driver/core/Host.java
@@ -36,6 +36,9 @@ public class Host {
     private volatile String datacenter;
     private volatile String rack;
 
+    private volatile boolean isUp;
+    private final ConvictionPolicy policy;
+
     // Tracks reconnection attempts to that host so we avoid adding multiple tasks
     final AtomicReference<ScheduledFuture<?>> reconnectionAttempt = new AtomicReference<ScheduledFuture<?>>();
 
@@ -48,7 +51,9 @@ public class Host {
             throw new NullPointerException();
 
         this.address = address;
-        this.monitor = new HealthMonitor(policy.create(this));
+        this.isUp = true;
+        this.policy = policy.create(this);
+        this.monitor = new HealthMonitor();
         this.defaultExecutionInfo = new ExecutionInfo(ImmutableList.of(this));
     }
 
@@ -95,6 +100,15 @@ public class Host {
     }
 
     /**
+     * Returns whether the host is considered up by the driver.
+     *
+     * @return whether the node is considered up.
+     */
+    public boolean isUp() {
+        return isUp;
+    }
+
+    /**
      * Returns the health monitor for this host.
      *
      * The health monitor keeps tracks of the known host state (up or down). A
@@ -102,7 +116,16 @@ public class Host {
      * the health monitor to be notified when this node is detected to be up or down
      *
      * @return the host {@link HealthMonitor}.
+     *
+     * @deprecated you are encouraged not to use the HealthMonitor anymore. To test
+     * if a node is considered UP or not, you should use {@link #isUp} instead. To
+     * register a {@link Host.StateListener}, you should do so at the Cluster level
+     * through {@link Cluster#register} (registering against the HealtMonitor does
+     * not work as intented: listeners will only be informed of onUp and onDown
+     * events (not onAdd and onRemove) and you need to manually register against
+     * every host. {@link Cluster#register} solves this).
      */
+    @Deprecated
     public HealthMonitor getMonitor() {
         return monitor;
     }
@@ -125,20 +148,42 @@ public class Host {
         return address.toString();
     }
 
+    void setDown() {
+        isUp = false;
+    }
+
+    /**
+     * Resets the policy, setting the host as up and informing the
+     * registered listener that the node is up.
+     */
+    void setUp() {
+        policy.reset();
+        isUp = true;
+        for (Host.StateListener listener : monitor.listeners)
+            listener.onUp(this);
+    }
+
+    boolean signalConnectionFailure(ConnectionException exception) {
+        boolean isDown = policy.addFailure(exception);
+        if (isDown) {
+            // For vague compatibility sake
+            for (Host.StateListener listener : monitor.listeners)
+                listener.onDown(this);
+        }
+        return isDown;
+    }
+
     /**
      * Tracks the health of a node and notify listeners when a host is considered up or down.
+     *
+     * @deprecated See {@link Host#getMonitor}.
      */
+    @Deprecated
     public class HealthMonitor {
 
-        private final ConvictionPolicy policy;
-
         private Set<Host.StateListener> listeners = new CopyOnWriteArraySet<Host.StateListener>();
-        private volatile boolean isUp;
 
-        HealthMonitor(ConvictionPolicy policy) {
-            this.policy = policy;
-            this.isUp = true;
-        }
+        HealthMonitor() {}
 
         /**
          * Registers the provided listener to be notified on up/down events.
@@ -146,7 +191,10 @@ public class Host {
          * Registering the same listener multiple times is a no-op.
          *
          * @param listener the new {@link Host.StateListener} to register.
+         *
+         * @deprecated See {@link Host#getMonitor}. Replaced by {@link Cluster#register}.
          */
+        @Deprecated
         public void register(StateListener listener) {
             listeners.add(listener);
         }
@@ -158,7 +206,10 @@ public class Host {
          * registered against this monitor.
          *
          * @param listener the {@link Host.StateListener} to unregister.
+         *
+         * @deprecated See {@link Host#getMonitor}. Replaced by {@link Cluster#unregister}.
          */
+        @Deprecated
         public void unregister(StateListener listener) {
             listeners.remove(listener);
         }
@@ -167,37 +218,12 @@ public class Host {
          * Returns whether the host is considered up by this monitor.
          *
          * @return whether the node is considered up.
+         *
+         * @deprecated See {@link Host#getMonitor}. Replaced by {@link Host#isUp}.
          */
+        @Deprecated
         public boolean isUp() {
-            return isUp;
-        }
-
-        void setDown() {
-            isUp = false;
-            for (Host.StateListener listener : listeners)
-                listener.onDown(Host.this);
-        }
-
-        /**
-         * Resets the monitor, setting the host as up and informing the
-         * registered listener that the node is up.
-         */
-        void setUp() {
-            policy.reset();
-            for (Host.StateListener listener : listeners)
-                listener.onUp(Host.this);
-            isUp = true;
-        }
-
-        boolean signalConnectionFailure(ConnectionException exception) {
-            // Already down, avoid duplicate signaling
-            if (!isUp)
-                return true;
-
-            boolean isDown = policy.addFailure(exception);
-            if (isDown)
-                setDown();
-            return isDown;
+            return Host.this.isUp;
         }
     }
 
@@ -214,7 +240,7 @@ public class Host {
 
         /**
          * Called when a new node is added to the cluster.
-         *
+         * <p>
          * The newly added node should be considered up.
          *
          * @param host the host that has been newly added.