+++ driver-core/src/test/java/com/datastax/driver/core/TestUtils.java
@@ -344,6 +344,6 @@ public abstract class TestUtils {
     }
 
     private static boolean testHost(Host host, boolean testForDown) {
-        return testForDown ? !host.getMonitor().isUp() : host.getMonitor().isUp();
+        return testForDown ? !host.isUp() : host.isUp();
     }
 }