+++ driver-core/CHANGELOG.rst
@@ -2,6 +2,11 @@ CHANGELOG
 =========
 
 * 1.0.2:
+  - [api] Host#getMonitor and all Host.HealthMonitor methods have been
+    deprecated. The new Host#isUp method is now prefered to the method
+    in the monitor and you should now register Host.StateListener against
+    the Cluster object directly (registering against a host HealthMonitor
+    was much more limited anyway).
   - [new] New serialize/deserialize methods in DataType to serialize/deserialize
     values to/from bytes (JAVA-92).
   - [new] New getIndexOf() method in ColumnDefinitions to find the index of